+++ crates/oberon/src/application.rs
@@ -9,5 +9,5 @@ pub trait ApplicationHandler
 {
     fn setup(&mut self, _config: &mut Config) {}
     fn before_start(&mut self, _canvas: Canvas<'_>) {}
-    fn frame(&mut self, canvas: Canvas<'_>, dt: f32, app_loop: &mut Arc<Loop>);
+    fn frame(&mut self, canvas: Canvas<'_>, dt: f64, app_loop: &mut Arc<Loop>);
 }
