+++ crates/oberon_core/src/linalg/point2f.rs
@@ -1,13 +1,13 @@
 #[derive(Copy, Clone, Debug)]
 pub struct Point2f
 {
-    pub x: f32,
-    pub y: f32,
+    pub x: f64,
+    pub y: f64,
 }
 
 impl Point2f
 {
-    pub const fn new(x: f32, y: f32) -> Self
+    pub const fn new(x: f64, y: f64) -> Self
     {
         Self { x, y }
     }
