+++ crates/oberon_core/src/linalg/shapes/mod.rs
@@ -7,5 +7,5 @@ pub trait Shape
 {
     fn area(&self) -> usize;
     fn points_outline(&self) -> impl Iterator<Item = Point2>;
-    fn rotate(&mut self, angle: f32);
+    fn rotate(&mut self, angle: f64);
 }
